Updated the OpenApi generate to use dotnet swagger
The Generation of the openapi.json file now uses the dotnet swagger tool.

For that to work a dependency on the `Swashbuckle.AspNetCore.Swagger` package has been added to allow that assembly to be used when running the `dotnet swagger` command.

The `Main()` function has been updated to account for when the entry assembly is "dotnet-swagger" and skip over the startup option parsing which would fail as swagger passes an `--applicationName` argument which is rejected by the command line parser.